+++ ee/api/hooks.py
@@ -1,4 +1,5 @@
 from typing import cast
+from urllib.parse import urlparse
 
 from django.conf import settings
 from rest_framework import exceptions, serializers, viewsets
@@ -21,6 +22,11 @@ class HookSerializer(serializers.ModelSerializer):
             raise exceptions.ValidationError(detail=f"Unexpected event {event}")
         return event
 
+    def validate_target(self, target):
+        if not valid_domain(target):
+            raise exceptions.ValidationError(detail=f"'hooks.zapier.com' is the only allowed target domain")
+        return target
+
 
 class HookViewSet(StructuredViewSetMixin, viewsets.ModelViewSet):
     """
@@ -35,3 +41,8 @@ class HookViewSet(StructuredViewSetMixin, viewsets.ModelViewSet):
     def perform_create(self, serializer):
         user = cast(User, self.request.user)
         serializer.save(user=user, team_id=self.team_id)
+
+
+def valid_domain(url) -> bool:
+    target_domain = urlparse(url).netloc
+    return bool(target_domain == "hooks.zapier.com")
