--- ee/tasks/test/test_hooks.py
@@ -1,19 +1,0 @@
-from ee.tasks.hooks import valid_domain
-
-
-def test_valid_domain() -> None:
-
-    test_cases = {
-        "http://hooks.zapier.com": True,
-        "https://hooks.zapier.com": True,
-        "http://hooks.zapier.com/something": True,
-        "https://hooks.zapier.com/something": True,
-        "http://hooks.zapierz.com": False,
-        "https://hooks.zapierz.com": False,
-        "http://hoos.zapier.com/something": False,
-        "https://hoos.zapier.com/something": False,
-    }
-
-    for test_input, expected_test_output in test_cases.items():
-        test_output = valid_domain(test_input)
-        assert test_output == expected_test_output
